abstract width/height logic into custom hook
